Make MpesaSecurity trait derivable
