+++ Cargo.toml
@@ -16,6 +16,9 @@ opt-level = 0
 [profile.release]
 opt-level = 3
 
+[workspace]
+members = ["mpesa_derive"]
+
 [dependencies]
 reqwest = { version = "0.10", features = ["blocking", "json"] }
 serde = { version = "1.0.114", features = ["derive"] }
@@ -23,3 +26,4 @@ serde_json = "1.0"
 dotenv = "0.15.0"
 base64 = "0.12.3"
 openssl = "0.10.30"
+mpesa_derive = { path = "./mpesa_derive" }
