+++ mpesa_derive/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "mpesa_derive"
+version = "0.1.0"
+authors = ["Collins Muriuki <murerwacollins@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "1.0"
+quote = "1.0"
+base64 = "0.12.3"
+openssl = "0.10.30"
