+++ src/client.rs
@@ -1,10 +1,7 @@
 use std::collections::HashMap;
-use std::error::Error;
 use reqwest::blocking::{Client, Response};
-use openssl::x509::X509;
-use openssl::rsa::Padding;
-use base64::encode;
 use serde_json::json;
+use mpesa_derive::MpesaSecurity;
 
 use super::environment::Environment;
 use crate::{CommandId, IdentifierTypes};
@@ -15,32 +12,12 @@ use crate::payloads::{AccountBalanceResponse,AccountBalancePayload};
 use crate::mpesa_security::MpesaSecurity;
 
 /// Mpesa client that will facilitate communication with the Safaricom API
-#[derive(Debug)]
+#[derive(Debug, MpesaSecurity)]
 pub struct Mpesa {
     client_key: String,
     client_secret: String,
     environment: Environment,
     initiator_password: String,
-}
-
-impl MpesaSecurity for Mpesa {
-    fn gen_security_credentials(&self) -> Result<String, Box<dyn Error>> {
-        let pem = self.environment.get_certificate().as_bytes();
-        let cert = X509::from_pem(pem).expect("error extracting X509 from pem");
-        // getting the public and rsa keys
-        let pub_key = cert.public_key().expect("error getting public key");
-        let rsa_key = pub_key.rsa().expect("error getting rsa key from pub_key");
-        // configuring the buffer
-        let buf_len = pub_key.size();
-        let mut buffer = vec![0; buf_len];
-
-        rsa_key.public_encrypt(
-            self.initiator_password.as_bytes(),
-            &mut buffer,
-            Padding::PKCS1,
-        )?;
-        Ok(encode(buffer))
-    }
 }
 
 impl Mpesa {
@@ -57,7 +34,7 @@ impl Mpesa {
     /// Generates an access token
     /// Sends `GET` request to Safaricom oauth to acquire token for token authentication
     /// The OAuth access token expires after an hour, after which, you will need to generate another access token
-    fn auth(&self) -> Result<String, Box<dyn Error>> {
+    fn auth(&self) -> Result<String, Box<dyn std::error::Error>> {
         let url = format!("{}/oauth/v1/generate?grant_type=client_credentials", self.environment.base_url());
 
         let resp: AuthResponse = Client::new().get(&url)
@@ -113,7 +90,7 @@ impl Mpesa {
         queue_timeout_url: &str,
         result_url: &str,
         occasion: &str
-    ) -> Result<B2cResponse, Box<dyn Error>> {
+    ) -> Result<B2cResponse, Box<dyn std::error::Error>> {
         let url = format!("{}/mpesa/b2c/v1/paymentrequest", self.environment.base_url());
         let credentials = self.gen_security_credentials()?;
 
@@ -200,7 +177,7 @@ impl Mpesa {
         queue_timeout_url: &str,
         result_url: &str,
         account_ref: &str,
-    ) -> Result<B2bResponse,Box<dyn Error>> {
+    ) -> Result<B2bResponse,Box<dyn std::error::Error>> {
         let url = format!("{}/mpesa/b2b/v1/paymentrequest", self.environment.base_url());
         let credentials = self.gen_security_credentials()?;
 
@@ -278,7 +255,7 @@ impl Mpesa {
         confirmation_url: &str,
         response_type: ResponseType,
         short_code: &str,
-    ) -> Result<Response, Box<dyn Error>> {
+    ) -> Result<Response, Box<dyn std::error::Error>> {
         let url = format!("{}/mpesa/c2b/v1/registerurl", self.environment.base_url());
 
         let payload = C2bRegisterPayload {
@@ -337,7 +314,7 @@ impl Mpesa {
         msisdn: &str,
         bill_ref_number: &str,
         short_code: &str,
-    ) -> Result<C2bSimulateResponse, Box<dyn Error>> {
+    ) -> Result<C2bSimulateResponse, Box<dyn std::error::Error>> {
         let url = format!("{}/mpesa/c2b/v1/simulate", self.environment.base_url());
 
         let payload = C2bSimulatePayload {
@@ -396,7 +373,7 @@ impl Mpesa {
         initiator_name: &str,
         queue_timeout_url: &str,
         result_url: &str,
-    ) -> Result<AccountBalanceResponse, Box<dyn Error>> {
+    ) -> Result<AccountBalanceResponse, Box<dyn std::error::Error>> {
         let url = format!("{}/mpesa/accountbalance/v1/query", self.environment.base_url());
         let credentials = self.gen_security_credentials()?;
