+++ src/mpesa_security.rs
@@ -1,3 +1,5 @@
+use std::error::Error;
+
 /// Trait responsible for implementation of security configs for Mpesa
 pub trait MpesaSecurity {
     /// Generates security credentials
@@ -5,3 +7,4 @@ pub trait MpesaSecurity {
     /// Security credentials are generated by encrypting the base64 encoded initiator password with M-Pesa’s public key, a X509 certificate.
     /// Returns base64 encoded string.
     fn gen_security_credentials(&self) -> Result<String, Box<dyn Error>>;
+}
